Move some DTOs to PNC-API project
