+++ core/src/test/java/org/jboss/pnc/causeway/ctl/ImportControllerTest.java
@@ -26,6 +26,13 @@ import com.redhat.red.build.koji.model.json.KojiImport;
 import com.redhat.red.build.koji.model.json.util.KojiObjectMapper;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
+import org.jboss.pnc.api.causeway.dto.CallbackTarget;
+import org.jboss.pnc.api.causeway.dto.push.Build;
+import org.jboss.pnc.api.causeway.dto.push.MavenBuild;
+import org.jboss.pnc.api.causeway.dto.push.MavenBuiltArtifact;
+import org.jboss.pnc.api.causeway.dto.push.NpmBuild;
+import org.jboss.pnc.api.causeway.dto.push.NpmBuiltArtifact;
+import org.jboss.pnc.api.dto.Request;
 import org.jboss.pnc.causeway.CausewayException;
 import org.jboss.pnc.causeway.CausewayFailure;
 import org.jboss.pnc.causeway.brewclient.BrewClient;
@@ -35,13 +42,6 @@ import org.jboss.pnc.causeway.config.CausewayConfig;
 import org.jboss.pnc.pncmetrics.MetricsConfiguration;
 import org.jboss.pnc.causeway.rest.BrewBuild;
 import org.jboss.pnc.causeway.rest.BrewNVR;
-import org.jboss.pnc.causeway.rest.CallbackMethod;
-import org.jboss.pnc.causeway.rest.CallbackTarget;
-import org.jboss.pnc.causeway.rest.model.Build;
-import org.jboss.pnc.causeway.rest.model.MavenBuild;
-import org.jboss.pnc.causeway.rest.model.MavenBuiltArtifact;
-import org.jboss.pnc.causeway.rest.model.NpmBuild;
-import org.jboss.pnc.causeway.rest.model.NpmBuiltArtifact;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -87,7 +87,7 @@ public class ImportControllerTest {
     private static final String ARTIFACTS_VERSION = "1.1.1.redhat_1";
     private static final String BUILD_NAME = "org.apache.geronimo.specs:geronimo-annotation_1.0_spec";
     private static final String CALLBACK_URL = "http://localhost:8081/callback";
-    private static final CallbackTarget CALLBACK_TARGET = new CallbackTarget(CALLBACK_URL, CallbackMethod.POST);
+    private static final CallbackTarget CALLBACK_TARGET = new CallbackTarget(CALLBACK_URL, Request.Method.POST);
     private static final String KOJI_URL = "http://koji.example.com/koji";
     private static final String KOJI_BUILD_URL = KOJI_URL + "/build?id=";
