--- rest/src/main/java/org/jboss/pnc/causeway/rest/model/Dependency.java
@@ -1,45 +1,0 @@
-/**
- * Copyright (C) 2015 Red Hat, Inc. (jbrazdil@redhat.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.causeway.rest.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.NonNull;
-
-/**
- *
- * @author Honza Brázdil &lt;janinko.g@gmail.com&gt;
- */
-@Data
-@Builder
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonDeserialize(builder = Dependency.DependencyBuilder.class)
-public class Dependency {
-    @NonNull
-    private final String filename;
-    @NonNull
-    private final String md5;
-    private final long size;
-
-    @JsonPOJOBuilder(withPrefix = "")
-    public static class DependencyBuilder {
-
-    }
-}
